tests/requirements-test.txt
pytest==7.4.3
pytest-cov==4.1.0
pytest-flask==1.3.0
selenium==4.15.2
webdriver-manager==4.0.1 